options/flags can be command specific
accepting flags which aren't supported by a command can be quiet
confusing especially if the name suggests that it could be supported.

While this adds the infrastructure this isn't completed yet, so many
commands still accept options that they shouldn't.

The direct effect in reality is that 'apt' is supporting the same commandline
flags as its apt-* counterparts would out-of-the-box now.