do not try to chown if not run as root
If this code is run as non-root we are in a special situation (e.g. in
our testcases) where it is obvious that we can't enforce user/group on
any file, so skip this code altogether instead of bugging users with
an error message.

The file is still handled with chmod, so all the security we can enforce
is still enforced of course.

Git-Dch: Ignore